+++ remotestate-ts/src/lib/react/provider.tsx
@@ -1,9 +1,6 @@
 import { useEffect, useMemo, type ReactNode } from "react";
-import {
-  createRemoteStateClient,
-  type RemoteStateClient,
-  type RemoteStateClientOptions,
-} from "../client";
+import { type RemoteStateClient, type RemoteStateClientOptions } from "../client";
+import { createRemoteStateClient } from "../remote";
 import { RemoteStateContext } from "./context";
 
 export interface RemoteStateProviderProps extends RemoteStateClientOptions {